Public key primitives moved to the separate module
