+++ parity-crypto/src/error.rs
@@ -20,22 +20,6 @@ use std::{fmt, result, error::Error as StdError};
 pub enum Error {
 	Scrypt(ScryptError),
 	Symm(SymmError),
-	/// Invalid secret key
-	InvalidSecret,
-	/// Invalid public key
-	InvalidPublic,
-	/// Invalid address
-	InvalidAddress,
-	/// Invalid EC signature
-	InvalidSignature,
-	/// Invalid AES message
-	InvalidMessage,
-	/// secp256k1 enc error
-	Secp(secp256k1::Error),
-	/// IO Error
-	Io(::std::io::Error),
-	/// Custom
-	Custom(String),
 }
 
 #[derive(Debug)]
@@ -63,9 +47,6 @@ impl StdError for Error {
 		match self {
 			Error::Scrypt(scrypt_err) => Some(scrypt_err),
 			Error::Symm(symm_err) => Some(symm_err),
-			Error::Secp(secp_err) => Some(secp_err),
-			Error::Io(err) => Some(err),
-			_ => None,
 		}
 	}
 }
@@ -95,14 +76,6 @@ impl fmt::Display for Error {
 		match self {
 			Error::Scrypt(err)=> write!(f, "scrypt error: {}", err),
 			Error::Symm(err) => write!(f, "symm error: {}", err),
-			Error::InvalidSecret => write!(f, "invalid secret"),
-			Error::InvalidPublic => write!(f, "invalid public"),
-			Error::InvalidAddress => write!(f, "invalid address"),
-			Error::InvalidSignature => write!(f, "invalid EC signature"),
-			Error::InvalidMessage => write!(f, "invalid AES message"),
-			Error::Secp(err) => write!(f, "secp error: {}", err),
-			Error::Io(err) => write!(f, "I/O error: {}", err),
-			Error::Custom(err) => write!(f, "custom crypto error: {}", err),
 		}
 	}
 }
@@ -128,32 +101,9 @@ impl fmt::Display for SymmError {
 	}
 }
 
-impl Into<String> for Error {
-	fn into(self) -> String {
-		format!("{}", self)
-	}
-}
-
 impl Into<std::io::Error> for Error {
 	fn into(self) -> std::io::Error {
 		std::io::Error::new(std::io::ErrorKind::Other, format!("Crypto error: {}",self))
-	}
-}
-
-impl From<::std::io::Error> for Error {
-	fn from(err: ::std::io::Error) -> Error {
-		Error::Io(err)
-	}
-}
-
-impl From<::secp256k1::Error> for Error {
-	fn from(e: ::secp256k1::Error) -> Error {
-		match e {
-			::secp256k1::Error::InvalidMessage => Error::InvalidMessage,
-			::secp256k1::Error::InvalidPublicKey => Error::InvalidPublic,
-			::secp256k1::Error::InvalidSecretKey => Error::InvalidSecret,
-			_ => Error::InvalidSignature,
-		}
 	}
 }
